Inline DataModelMatchers into typed matcher classes
DataModelMatchers was an aggregator that delegated to typed matchers
via their inner classes. Move those factory methods to their proper
typed matcher classes and update all callers to import directly.

- Add public factory methods to ConvertConfigurationAnswerElementMatchers,
  ParseVendorConfigurationAnswerElementMatchers, PermittedByAclMatchers,
  and ZoneMatchers (previously package-private with no factories).
- Add hasOutgoingFilterName to InterfaceMatchers and
  hasReferenceBandwidth(double) to OspfProcessMatchers.
- Update 20 test files to import from typed matchers directly.
- Reduce DataModelMatchers to just the generic forAll() utility.

----

Prompt:
```
Should some of these DataModelMatchers that are just one-liners
delegating to a different matcher just be deleted via inlining?
```

commit-id:b093a646
