+++ projects/batfish/src/test/java/org/batfish/job/ConvertConfigurationJobTest.java
@@ -2,7 +2,7 @@ package org.batfish.job;
 
 import static org.batfish.common.matchers.WarningMatchers.hasText;
 import static org.batfish.datamodel.Configuration.DEFAULT_VRF_NAME;
-import static org.batfish.datamodel.matchers.DataModelMatchers.hasDefinedStructureWithDefinitionLines;
+import static org.batfish.datamodel.matchers.ConvertConfigurationAnswerElementMatchers.hasDefinedStructureWithDefinitionLines;
 import static org.batfish.datamodel.matchers.MapMatchers.hasKeys;
 import static org.batfish.datamodel.matchers.StaticRouteMatchers.hasTrack;
 import static org.batfish.datamodel.tracking.TrackMethods.alwaysTrue;
