+++ projects/batfish/src/test/java/org/batfish/representation/aws/SubnetTest.java
@@ -6,10 +6,10 @@ import static org.batfish.datamodel.Configuration.DEFAULT_VRF_NAME;
 import static org.batfish.datamodel.Interface.NULL_INTERFACE_NAME;
 import static org.batfish.datamodel.matchers.ConfigurationMatchers.hasDeviceModel;
 import static org.batfish.datamodel.matchers.ConfigurationMatchers.hasVrf;
-import static org.batfish.datamodel.matchers.DataModelMatchers.hasIncomingFilter;
-import static org.batfish.datamodel.matchers.DataModelMatchers.hasOutgoingFilter;
 import static org.batfish.datamodel.matchers.InterfaceMatchers.hasFirewallSessionInterfaceInfo;
+import static org.batfish.datamodel.matchers.InterfaceMatchers.hasIncomingFilter;
 import static org.batfish.datamodel.matchers.InterfaceMatchers.hasName;
+import static org.batfish.datamodel.matchers.InterfaceMatchers.hasOutgoingFilter;
 import static org.batfish.datamodel.matchers.InterfaceMatchers.hasVrfName;
 import static org.batfish.datamodel.matchers.VrfMatchers.hasStaticRoutes;
 import static org.batfish.representation.aws.AwsConfigurationTestUtils.getTestSubnet;
