+++ projects/common/src/test/java/org/batfish/datamodel/matchers/PermittedByAclMatchers.java
@@ -6,9 +6,23 @@ import org.batfish.datamodel.acl.PermittedByAcl;
 import org.hamcrest.FeatureMatcher;
 import org.hamcrest.Matcher;
 
-final class PermittedByAclMatchers {
+public final class PermittedByAclMatchers {
 
-  static class HasAclName extends FeatureMatcher<PermittedByAcl, String> {
+  public static @Nonnull Matcher<PermittedByAcl> hasAclName(
+      @Nonnull Matcher<? super String> subMatcher) {
+    return new HasAclName(subMatcher);
+  }
+
+  public static @Nonnull Matcher<PermittedByAcl> hasAclName(@Nonnull String name) {
+    return hasAclName(org.hamcrest.Matchers.equalTo(name));
+  }
+
+  public static @Nonnull Matcher<AclLineMatchExpr> isPermittedByAclThat(
+      @Nonnull Matcher<? super PermittedByAcl> subMatcher) {
+    return new IsPermittedByAclThat(subMatcher);
+  }
+
+  private static class HasAclName extends FeatureMatcher<PermittedByAcl, String> {
 
     public HasAclName(@Nonnull Matcher<? super String> subMatcher) {
       super(subMatcher, "A PermittedByAcl with aclName:", "aclName");
@@ -20,7 +34,8 @@ final class PermittedByAclMatchers {
     }
   }
 
-  static class IsPermittedByAclThat extends IsInstanceThat<AclLineMatchExpr, PermittedByAcl> {
+  private static final class IsPermittedByAclThat
+      extends IsInstanceThat<AclLineMatchExpr, PermittedByAcl> {
     IsPermittedByAclThat(@Nonnull Matcher<? super PermittedByAcl> subMatcher) {
       super(PermittedByAcl.class, subMatcher);
     }
