+++ projects/common/src/test/java/org/batfish/datamodel/matchers/ZoneMatchers.java
@@ -1,13 +1,19 @@
 package org.batfish.datamodel.matchers;
 
 import java.util.SortedSet;
+import javax.annotation.Nonnull;
 import org.batfish.datamodel.Zone;
 import org.hamcrest.FeatureMatcher;
 import org.hamcrest.Matcher;
 
-final class ZoneMatchers {
+public final class ZoneMatchers {
 
-  static final class HasInterfaces extends FeatureMatcher<Zone, SortedSet<String>> {
+  public static @Nonnull Matcher<Zone> hasMemberInterfaces(
+      @Nonnull Matcher<? super SortedSet<String>> subMatcher) {
+    return new HasInterfaces(subMatcher);
+  }
+
+  private static final class HasInterfaces extends FeatureMatcher<Zone, SortedSet<String>> {
 
     public HasInterfaces(Matcher<? super SortedSet<String>> subMatcher) {
       super(subMatcher, "A zone with interfaces:", "interfaces");
